Replaced wildcard parameter * by ~
Otherwise, some strings may be interpreted as regular expression. The
correct syntax for selecting variables with regular expression is:

o{'[\w*_US]'}

while

o{'\w*_US'}

is not correct. The user must explicitely state that s/he wants to
select variables with regular expression by beginning and closing the
string with square brackets.

When a wildcard parameter is detected, the string is converted to a
regular expression by adding opening and closing square brackets and
replacing the wildcard parameter by \w*.

Consequently, in the last example, if * is the wildcard parameter,
string '\w*_US' would be converted into '[\w\w*_US]'. That is the
reason why this commit changes the wildcard parameter to ~ (which is
not used in matlab's regular expressions).